move validate_source_Sv to echopype/utils/data_proc_lvls.py and create a simple test for it
